Renamed custom action projects, added build process for DLLs
