+++ installer/win/BracketsConfigurator/BracketsConfigurator/BracketsConfigurator.config
