+++ installer/win/LaunchBrackets/LaunchBrackets/LaunchBrackets.config
