+++ installer/win/LaunchBrackets/LaunchBrackets/LaunchBrackets.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{4922460C-3623-4C0D-900B-3E69E8F1EE98}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>CustomAction2</RootNamespace>
-    <AssemblyName>CustomAction2</AssemblyName>
+    <RootNamespace>LaunchBrackets</RootNamespace>
+    <AssemblyName>LaunchBrackets</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -41,14 +41,13 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CustomAction.cs" />
+    <Compile Include="LaunchBrackets.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Content Include="CustomAction.config" />
+    <Content Include="LaunchBrackets.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(WixCATargetsPath)" Condition=" '$(WixCATargetsPath)' != '' " />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\WiX\v3.x\Wix.CA.targets" Condition=" '$(WixCATargetsPath)' == '' AND Exists('$(MSBuildExtensionsPath32)\Microsoft\WiX\v3.x\Wix.CA.targets') " />
-  <Target Name="EnsureWixToolsetInstalled" Condition=" '$(WixCATargetsImported)' != 'true' ">
-    <Error Text="The WiX Toolset v3.11 (or newer) build tools must be installed to build this project. To download the WiX Toolset, see http://wixtoolset.org/releases/" />
+  <Target Name="EnsureWixToolsetInstalled">
   </Target>
 </Project>
