+++ installer/win/LaunchBrackets/LaunchBracketsSolution.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LaunchBrackets", "LaunchBrackets\LaunchBrackets.csproj", "{4922460C-3623-4C0D-900B-3E69E8F1EE98}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x86 = Debug|x86
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4922460C-3623-4C0D-900B-3E69E8F1EE98}.Debug|x86.ActiveCfg = Debug|x86
+		{4922460C-3623-4C0D-900B-3E69E8F1EE98}.Debug|x86.Build.0 = Debug|x86
+		{4922460C-3623-4C0D-900B-3E69E8F1EE98}.Release|x86.ActiveCfg = Release|x86
+		{4922460C-3623-4C0D-900B-3E69E8F1EE98}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
